requirements.txt
google-genai>=1.0
pillow>=11.3.0
python-dotenv>=1.0.0
numpy
pydantic